Correcciones chabooooooooooonessssss

+++ IOO/Codigo/codigo/src/ar/edu/uade/ioo/grupo5/tpo/bo/Producto.java
@@ -59,7 +59,7 @@ public class Producto {
 		return proveedor;
 	}
 
-	public boolean tieneStockDisponible(int cantidad) {
+	public boolean tieneStockDisponible(double cantidad) {
 		return (stock - cantidad) >= 0;
 
 	}
